Add quotes module that uses yfinance for downloading quotes. Refactor models into common/transactions.
