+++ investd/__main__.py
@@ -5,13 +5,17 @@ import click
 
 from . import reports
 from .config import PERSIST_PATH
+from .quotes import generate_quotes_csv
 from .sources import ingest_sources_as_df
 
 app_name = "investd"
 log = logging.getLogger(app_name)
 
 
-@click.command(name="ingest-sources")
+cli = click.Group(name=app_name, help="investd - A tool for summarizing investments.")
+
+
+@cli.command(name="ingest-sources")
 @click.option(
     "--output",
     type=click.Path(dir_okay=False, writable=True),
@@ -24,7 +28,7 @@ def ingest_sources_cmd(output: Path):
     df_tx.to_csv(output, index=False)
 
 
-@click.command(name="report")
+@cli.command(name="report")
 @click.option(
     "--report",
     default="overview",
@@ -45,11 +49,10 @@ def report_cmd(report: str, ingest: bool):
     log.info(f"Report created: {path_output}")
 
 
-cli = click.Group(
-    name=app_name,
-    help="investd - A tool for summarizing investments.",
-    commands=[ingest_sources_cmd, report_cmd],
-)
+@cli.command(name="download-quotes")
+def quotes_cmd():
+    generate_quotes_csv()
+
 
 if __name__ == "__main__":
     cli()
