+++ investd/sources/xtb.py
@@ -4,7 +4,8 @@ from typing import Iterator
 
 import pandas as pd
 
-from ..model import Action, AssetType, Currency, Transaction
+from ..common import Action, AssetType, Currency
+from ..transactions import Transaction
 from .base import SourceBase
