+++ setup.cfg
@@ -1,7 +1,10 @@
 [mypy]
 plugins = pydantic.mypy
-ignore_missing_imports = true
+ignore_missing_imports = false
 follow_imports = silent
+
+[mypy-pandas]
+ignore_missing_imports = True
 
 [isort]
 profile = black
