+++ tests/sources/test_revolut_stocks.py
@@ -2,8 +2,9 @@ from datetime import datetime
 
 from pytest import approx
 
-from investd.model import Action, AssetType, Currency, Transaction
+from investd.common import Action, AssetType, Currency
 from investd.sources.revolut_stocks import RevolutStocks
+from investd.transactions import Transaction
 
 
 def test_parse_revolut_stocks(path_revolut_csv):
