+++ tests/sources/test_xtb.py
@@ -2,8 +2,9 @@ from datetime import datetime
 
 from pytest import approx
 
-from investd.model import Action, AssetType, Currency, Transaction
+from investd.common import Action, AssetType, Currency
 from investd.sources.xtb import XTB
+from investd.transactions import Transaction
 
 
 def test_parse_xtb(path_xtb_xlsx):
